requirements.txt
streamlit>=1.30.0
openai>=1.0.0
PyGithub>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
